refactor(DEEP-BAU): Change package structure
