+++ src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/config/MongoConfig.java
@@ -16,12 +16,12 @@ import org.springframework.data.auditing.DateTimeProvider;
 import org.springframework.data.mongodb.config.EnableMongoAuditing;
 import org.springframework.data.mongodb.core.convert.MongoCustomConversions;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-import uk.gov.companieshouse.chs.gov.uk.notify.integration.api.mongo.converter.DateToOffsetDateTimeConverter;
-import uk.gov.companieshouse.chs.gov.uk.notify.integration.api.mongo.converter.DocumentToSendEmailResponseConverter;
-import uk.gov.companieshouse.chs.gov.uk.notify.integration.api.mongo.converter.OffsetDateTimeToDateConverter;
+import uk.gov.companieshouse.chs.gov.uk.notify.integration.api.converter.DateToOffsetDateTimeConverter;
+import uk.gov.companieshouse.chs.gov.uk.notify.integration.api.converter.DocumentToSendEmailResponseConverter;
+import uk.gov.companieshouse.chs.gov.uk.notify.integration.api.converter.OffsetDateTimeToDateConverter;
 
 @Configuration
-@EnableMongoRepositories("uk.gov.companieshouse.chs.gov.uk.notify.integration.api.mongo.repository")
+@EnableMongoRepositories("uk.gov.companieshouse.chs.gov.uk.notify.integration.api.dao")
 @EnableMongoAuditing(dateTimeProviderRef = "mongodbDatetimeProvider")
 public class MongoConfig {
