--- src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/config/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Application wide configuration details
- */
-package uk.gov.companieshouse.chs.gov.uk.notify.integration.api.config;
