+++ src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/converter/DateToOffsetDateTimeConverter.java
@@ -1,0 +1,14 @@
+package uk.gov.companieshouse.chs.gov.uk.notify.integration.api.converter;
+
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
+import java.util.Date;
+
+import org.springframework.core.convert.converter.Converter;
+
+public class DateToOffsetDateTimeConverter implements Converter<Date, OffsetDateTime> {
+    @Override
+    public OffsetDateTime convert(Date source) {
+        return source.toInstant().atOffset(ZoneOffset.UTC);
+    }
+}
