+++ src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/converter/OffsetDateTimeToDateConverter.java
@@ -1,0 +1,13 @@
+package uk.gov.companieshouse.chs.gov.uk.notify.integration.api.converter;
+
+import java.time.OffsetDateTime;
+import java.util.Date;
+
+import org.springframework.core.convert.converter.Converter;
+
+public class OffsetDateTimeToDateConverter implements Converter<OffsetDateTime, Date> {
+    @Override
+    public Date convert(OffsetDateTime source) {
+        return Date.from(source.toInstant());
+    }
+}
