--- src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/lettergovuknotifypayload/package-info.java
@@ -1,4 +1,0 @@
-/**
- * generate the payload to be sent to the Gob.uk Notify send letter endpoint
- */
-package uk.gov.companieshouse.chs.gov.uk.notify.integration.api.lettergovuknotifypayload;
