+++ src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/model/NotificationLetterRequest.java
@@ -1,0 +1,12 @@
+package uk.gov.companieshouse.chs.gov.uk.notify.integration.api.model;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+import org.springframework.data.mongodb.core.mapping.Field;
+import uk.gov.companieshouse.api.chs_gov_uk_notify_integration_api.model.GovUkLetterDetailsRequest;
+
+@Document(collection = "letter_details")
+public record NotificationLetterRequest(
+        @Id String id,
+        @Field("request") GovUkLetterDetailsRequest request
+) {}
