--- src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/pdfgenerator/package-info.java
@@ -1,4 +1,0 @@
-/**
- * The public interface for the pdf generator component
- */
-package uk.gov.companieshouse.chs.gov.uk.notify.integration.api.pdfgenerator;
