--- src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/restapi/package-info.java
@@ -1,4 +1,0 @@
-/**
- * The REST API listener for the /letter and /email endpoints
- */
-package uk.gov.companieshouse.chs.gov.uk.notify.integration.api.restapi;
