--- src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/templatelookup/package-info.java
@@ -1,4 +1,0 @@
-/**
- * lookup the correct template based on the request details
- */
-package uk.gov.companieshouse.chs.gov.uk.notify.integration.api.templatelookup;
