--- src/main/java/uk/gov/companieshouse/chs/gov/uk/notify/integration/api/templatepersonalisation/package-info.java
@@ -1,4 +1,0 @@
-/**
- * replace template placeholders with personalisation details
- */
-package uk.gov.companieshouse.chs.gov.uk.notify.integration.api.templatepersonalisation;
